+++ pkg/apis/backup/v1/backup_spec.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -35,6 +35,9 @@ type ArangoBackupSpec struct {
 	PolicyName *string `json:"policyName,omitempty"`
 
 	Backoff *ArangoBackupSpecBackOff `json:"backoff,omitempty"`
+
+	// Lifetime is the time after which the backup will be deleted. Format: "1.5h" or "2h45m".
+	Lifetime string `json:"lifetime,omitempty"`
 }
 
 type ArangoBackupSpecDeployment struct {