+++ pkg/apis/backup/v1/backup_validate.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,7 +20,11 @@
 
 package v1
 
-import "github.com/arangodb/kube-arangodb/pkg/util/errors"
+import (
+	"time"
+
+	"github.com/arangodb/kube-arangodb/pkg/util/errors"
+)
 
 func (a *ArangoBackup) Validate() error {
 	if err := a.Spec.Validate(); err != nil {
@@ -48,6 +52,13 @@ func (a *ArangoBackupSpec) Validate() error {
 	if a.Upload != nil {
 		if err := a.Upload.Validate(); err != nil {
 			return err
+		}
+	}
+
+	if a.Lifetime != "" {
+		_, err := time.ParseDuration(a.Lifetime)
+		if err != nil {
+			return errors.Newf("Invalid duration: %v\n", err)
 		}
 	}